Laravel Project

+++ Laravel-Project/app/Http/Controllers/BookController.php
@@ -8,12 +8,20 @@ use App\Models\User;
 use App\Models\BookAuthor;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class BookController extends Controller
 {
     public function index(Request $request)
     {
-        $query = User::find(auth('web')->user()->id)->with('authors');
+        $user = auth('web')->user();
+        if ($user->role == 1) {
+            $author = $user->author;
+            $query = $author->books();
+        } else if ($user->role == 2) {
+            $query = Book::with('authors');
+        }
+       /* $query = Author::with('books')->find($author_id)->books();*/
         if ($request->input('search_title')) {
             $query = $query->where('title', 'like', '%' . $request->input('search_title') . '%');
         }
@@ -29,19 +37,29 @@ class BookController extends Controller
 
     public function store(Request $request)
     {
+        $user = auth('web')->user();
         $book = new Book;
         $book->fill(array('title' => $request->input('title'), 'price' => $request->input('price')));
-        $this->validate($request, [
+        $validBook = [
             'title' => 'required|min:8',
             'price' => 'required|numeric',
-            'authors' => 'required'
-        ]);
+        ];
+        if ($user->role == 2){
+            $validBook['authors'] = "required";
+        }
+        $this->validate($request, $validBook);
         $book->save();
         $book_id = $book->id;
-        foreach ($request->input('authors') as $author) {
+        if ($user->role == 1) {
             $bookAuthor = new BookAuthor;
-            $bookAuthor->fill(array('author_id' => $author, 'book_id' => $book_id));
+            $bookAuthor->fill(array('author_id' => $user->id, 'book_id' => $book_id));
             $bookAuthor->save();
+        } else {
+            foreach ($request->input('authors') as $author) {
+                $bookAuthor = new BookAuthor;
+                $bookAuthor->fill(array('author_id' => $author, 'book_id' => $book_id));
+                $bookAuthor->save();
+            }
         }
         return redirect()->route('books.create')->with('message', 'Գիրքը հաջողությամբ ստեղծվել է');
     }
@@ -62,29 +80,35 @@ class BookController extends Controller
 
     public function update(Request $request, $id)
     {
+        $user = auth('web')->user();
         $book = Book::with('booksAuthors')->find($id);
         $book->title = $request->input('title');
         $book->price = $request->input('price');
-        $this->validate($request, [
+        $validBook = [
             'title' => 'required|min:8',
             'price' => 'required|numeric',
-            'authors' => 'required'
-        ]);
-        $authors = $request->input('authors');
-        $booksAuthorsId = [];
-        foreach ($book->booksAuthors as $author_books) {
-            array_push($booksAuthorsId, $author_books->author_id);
+        ];
+        if ($user->role == 2) {
+            $validBook['authors'] = 'required';
         }
-        $deletedRepeatOne = array_diff($booksAuthorsId, $authors);
-        $deletedRepeatTwo = array_diff($authors, $booksAuthorsId);
-        if (!empty($deletedRepeatOne)) {
-            BookAuthor::where('book_id', $id)->whereIn('author_id', $deletedRepeatOne)->delete();
-        }
-        if (!empty($deletedRepeatTwo)) {
-            foreach ($deletedRepeatTwo as $author) {
-                $bookAuthor = new BookAuthor;
-                $bookAuthor->fill(array('author_id' => $author, 'book_id' => $id));
-                $bookAuthor->save();
+        $this->validate($request, $validBook);
+        if ($user->role == 2) {
+            $authors = $request->input('authors');
+            $booksAuthorsId = [];
+            foreach ($book->booksAuthors as $author_books) {
+                array_push($booksAuthorsId, $author_books->author_id);
+            }
+            $deletedRepeatOne = array_diff($booksAuthorsId, $authors);
+            $deletedRepeatTwo = array_diff($authors, $booksAuthorsId);
+            if (!empty($deletedRepeatOne)) {
+                BookAuthor::where('book_id', $id)->whereIn('author_id', $deletedRepeatOne)->delete();
+            }
+            if (!empty($deletedRepeatTwo)) {
+                foreach ($deletedRepeatTwo as $author) {
+                    $bookAuthor = new BookAuthor;
+                    $bookAuthor->fill(array('author_id' => $author, 'book_id' => $id));
+                    $bookAuthor->save();
+                }
             }
         }
         $book->save();
